requirements.txt
python_dotenv
dashscope
dashscope[generation]
langchain